refactor(DataTable): move ManageColumnsPopover into ManageColumns component
